Split picking to atomic functions
