+++ src/commands/pick.ts
@@ -1,7 +1,7 @@
 import { Command } from '../classes'
 import { CommandNames } from '../constants'
 import Log from '../utils/logger'
-import { randomPicker } from '../utils/randomPicker'
+import { picker } from '../utils/picker'
 
 const DEFAULT_MESSAGE = 'I choose you!'
 
@@ -13,7 +13,7 @@ export default new Command(
   async ({ ack, command: { channel_id }, respond }, slackAppInstance) => {
     await ack()
     await respond('Aye aye captain! Wait, I am the captain...')
-    await randomPicker({ channelId: channel_id, message: DEFAULT_MESSAGE }, slackAppInstance).catch(
+    await picker({ channelId: channel_id, message: DEFAULT_MESSAGE }, slackAppInstance).catch(
       async (error) => {
         await respond('Sorry. There was some kind of arr-or. :pirate_flag:')
         Log.error(error)
