+++ src/commands/setup.ts
@@ -5,8 +5,8 @@ import { Command } from '../classes'
 import ScheduledJobs from '../classes/ScheduledJobs'
 import { BlockIds, CommandNames } from '../constants'
 import { scheduleCronJob } from '../utils/cron'
-import { sendMessage } from '../utils/helpers'
 import Log from '../utils/logger'
+import { sendMessage } from '../utils/messages'
 import { getSetupModal } from '../utils/modals/modalGenerators'
 import { Prisma } from '../utils/prismaClient'
