+++ src/utils/picker/utils/getRandomArrayElements.ts
@@ -1,0 +1,4 @@
+export const getRandomArrayElements = <T>(array: T[], count: number) => {
+  const shuffled = array.sort(() => Math.random() - 0.5)
+  return shuffled.slice(0, count)
+}
